+++ CHANGELOG.md
@@ -14,6 +14,7 @@ title: Changelog
 * `<cc-input-text>`: remove "clipboard-copy" dependency
 * refactor: use `??` instead of `||` when it makes more sense
 * refactor: try to use `?.` when it's a bit simpler
+* i18n: use Intl.PluralRules instead of custom code
 
 ## 7.0.0 (2021-09-03)